moving out flight window into separate file
